+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0](https://github.com/microsoft/kiota-php/compare/microsoft-kiota-abstractions-v2.0.2...microsoft-kiota-abstractions-v2.1.0) (2026-08-27)
+
+
+### Features
+
+* adds support for wildcard host validation ([#102](https://github.com/microsoft/kiota-php/issues/102)) ([ec2b741](https://github.com/microsoft/kiota-php/commit/ec2b7410ff5a13396292d53c191a954b4623d7d2))
+* Handle map-style query parameters with null values ([d848f20](https://github.com/microsoft/kiota-php/commit/d848f20d7de5afecee7ac80e3c39732a6c1b85d3))
+
 ## [2.0.2](https://github.com/microsoft/kiota-php/compare/microsoft-kiota-abstractions-v2.0.1...microsoft-kiota-abstractions-v2.0.2) (2026-04-20)